landnotev3/pyproject.toml
[project]
name = "landnote-crawler"
version = "0.1.0"
description = "A unified crawler for landnote real estate and law exam data."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "beautifulsoup4>=4.12.0",
    "pikepdf>=8.0.0",
    "tqdm>=4.66.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "fake-useragent>=1.4.0",
    "urllib3>=2.0.0",
    "python-dotenv>=1.0.0",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.23.0",
    "reportlab>=4.0.0"
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []
